+++ src/js/control-bar/progress-control/play-progress-bar.js
@@ -2,6 +2,8 @@
  * @file play-progress-bar.js
  */
 import Component from '../../component.js';
+import * as Fn from '../../utils/fn.js';
+import formatTime from '../../utils/format-time.js';
 
 /**
  * Shows play progress
@@ -12,6 +14,12 @@ import Component from '../../component.js';
  * @class PlayProgressBar
  */
 class PlayProgressBar extends Component {
+
+  constructor(player, options){
+    super(player, options);
+    this.on(player, 'timeupdate', this.updateDataAttr);
+    player.ready(Fn.bind(this, this.updateDataAttr));
+  }
 
   /**
    * Create the component's DOM element
@@ -24,6 +32,11 @@ class PlayProgressBar extends Component {
       className: 'vjs-play-progress',
       innerHTML: `<span class="vjs-control-text"><span>${this.localize('Progress')}</span>: 0%</span>`
     });
+  }
+
+  updateDataAttr() {
+    let time = (this.player_.scrubbing()) ? this.player_.getCache().currentTime : this.player_.currentTime();
+    this.el_.setAttribute('data-current-time', formatTime(time, this.player_.duration()));
   }
 
 }